pyproject.toml
[project]
name = "game-recipe-book"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.4",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "uvicorn>=0.38.0",
    "python-multipart>=0.0.9",
    "openai>=1.6.1",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.0",
    "bcrypt>=4.0.0",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
pythonVersion = "3.12"
